Przeniesienie funkcji usera do osobnego pliku
