+++ Biblioteka-projektIO.cpp
@@ -9,134 +9,11 @@
 #include "Book.h"
 #include "User.h"
 #include "Librarian.h"
+#include "UserFunctions.h"
 
 using namespace std;
 
-void loading();
-
-void menuUser(User & registeredUser);
-
-void userLogin(vector <User>& userDatabase);
-
-void registration(vector<User>& userDatabase);
-
 void start(vector <User>& userDatabase);
-
-void loading() //animacja ladowania
-{
-	for (int i = 0; i < 3; ++i) //powtorz 3 razy
-	{
-		Sleep(200); //zaczekaj 200 jednostek
-		cout << "*" << flush; //wyswietl jedna gwiazdke
-		Sleep(200); //zaczekaj
-		cout << "*" << flush; //wyswietl
-		Sleep(200);
-		cout << "*" << flush;
-		Sleep(200);
-		cout << "*" << flush;
-		Sleep(200);
-		cout << "*" << flush;
-		Sleep(200);
-		cout << "\b\b\b\b\b     \b\b\b\b\b" << flush; //wymaz wszystkie gwiazdki
-	}
-	cout << "\n";
-} 
-
-void menuUser(User & registeredUser)
-{
-	cout << "Witaj " << registeredUser.getName() << "Wybierz jedna z dostepnych opcji wpisujac jej numer:\n";
-	cout << "\t1. Przegladaj ksiazki\n";
-	cout << "\t2. Wyszukaj ksiazke\n";
-}
-
-void userLogin(vector<User>& userDatabase)
-{
-	system("cls");
-	cout << "Logowanie\n\n";
-	string login;
-	string pass;
-	cout << "Login: ";
-	cin >> login;
-	for (auto i : userDatabase)
-	{
-		if (i.getName() == login)
-		{
-			cout << "Haslo: ";
-			cin >> pass;
-			while (pass != i.getPassword())
-			{
-				cout << "Haslo niepoprawne!\nSprobuj ponownie\n";
-				system("pause");
-				cout << "Logowanie\n\nLogin: " << login << "Haslo: ";
-				cin >> pass;
-			}
-			if (i.getPassword() == pass)
-			{
-				cout << "Logowanie poprawne!\n";
-				system("pause");
-				system("cls");
-				menuUser(i);
-			}
-		}
-	}
-	return;
-}
-
-void registration(vector<User>& userDatabase)
-{
-	system("cls");
-	string login;
-	cout << "Rejestracja\n\nPodaj login: ";
-	cin >> login;
-	bool loginAvailable = true;
-	do
-	{
-		for (auto i : userDatabase)
-		{
-			if (i.getName() == login) //jezeli login zajety
-			{
-				loginAvailable = false; //login niedostepny
-				cout << "Login zajety! wybierz inny!\n";
-				system("pause");
-				system("cls");
-				cout << "Rejestracja\n\nPodaj login: "; //wyswietl wszystko od nowa;
-				cin >> login;
-				break;
-			}
-			loginAvailable = true; //jesli nie wybrejkowalismy sie z petli for wczesniej, to znaczy ze login ok
-		}
-	} while (!loginAvailable);
-	string password;
-	cout << "Login poprawny. \nPodaj haslo: ";
-	cin >> password;
-	bool goodPassword = false;
-	while (!goodPassword)
-	{
-		if (password.size() > 8)
-		{
-			goodPassword = true;
-			break;
-		}
-		else
-		{
-			cout << "Haslo za slabe! podaj inne\n";
-			system("pause");
-			system("cls");
-			cout << "Rejestracja\n\nPodaj login :" << login << "\nLogin poprawny.\nPodaj haslo: "; //wyswietl wszystko od nowa;"
-			cin >> password;
-		}
-	}
-	
-	cout << "Haslo silne\nTworzenie konta";
-	loading();
-	User newUser(login, password, {1, 1, 1970});
-	cout << "Uzytkownik zostal utowrzony ";
-	userDatabase.push_back(newUser);
-	cout << " i dodany do bazy.\nDane uzytkownika:\n";
-	newUser.printUser();
-	system("pause");
-	start(userDatabase);
-}
 
 void start(vector <User>& userDatabase)
 {
@@ -166,7 +43,7 @@ void start(vector <User>& userDatabase)
 		}
 		case 4:
 		{
-			registration(userDatabase);
+			userRegistration(userDatabase);
 			break;
 		}
 		default:
